wip - api client, list moves
