+++ common/lib/api-client.js
@@ -34,28 +34,16 @@ function defineModels (jsonApi) {
     date_of_birth: '',
     gender: {
       jsonApi: 'hasOne',
-      type: 'gender',
+      type: 'genders',
     },
     ethnicity: {
       jsonApi: 'hasOne',
-      type: 'ethnicity',
+      type: 'ethnicities',
     },
-    identifiers: {
-      jsonApi: 'hasMany',
-      type: 'profile_identifier',
-    },
-    risk_alerts: {
-      jsonApi: 'hasMany',
-      type: 'profile_attribute',
-    },
-    health_alerts: {
-      jsonApi: 'hasMany',
-      type: 'profile_attribute',
-    },
-    court_information: {
-      jsonApi: 'hasMany',
-      type: 'profile_attribute',
-    },
+    identifiers: '',
+    risk_alerts: '',
+    health_alerts: '',
+    court_information: '',
   })
 
   jsonApi.define('gender', {
@@ -67,18 +55,6 @@ function defineModels (jsonApi) {
     code: '',
     title: '',
     description: '',
-  })
-
-  jsonApi.define('profile_identifier', {
-    value: '',
-    identifier_type: '',
-  })
-
-  jsonApi.define('profile_attribute', {
-    date: '',
-    expiry_date: '',
-    description: '',
-    comments: '',
   })
 
   jsonApi.define('location', {
